refactor: moved LocationServices to services
and moved all location services part to this class instead of methods in WelcomePage
