--- Notify/Notify/Notify/LocationServices.cs
@@ -1,54 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Xamarin.Essentials;
-using Xamarin.Forms;
-
-namespace Notify
-{
-    public class LocationService
-    {
-        readonly bool stopping = false;
-        public LocationService()
-        {
-        }
-
-        public async Task Run(CancellationToken token)
-        {
-            await Task.Run(async () => {
-                while (!stopping)
-                {
-                    token.ThrowIfCancellationRequested();
-                    try
-                    {
-                        await Task.Delay(2000);
-
-                        var request = new GeolocationRequest(GeolocationAccuracy.High);
-                        var location = await Geolocation.GetLocationAsync(request, token);
-                        if (location != null)
-                        {
-                            var message = new LocationMessage
-                            {
-                                Latitude = location.Latitude,
-                                Longitude = location.Longitude
-                            };
-
-                            Device.BeginInvokeOnMainThread(() =>
-                            {
-                                MessagingCenter.Send(message, "Location");
-                            });
-                        }
-                    }
-                    catch (Exception)
-                    {
-                        Device.BeginInvokeOnMainThread(() =>
-                        {
-                            var errormessage = new LocationErrorMessage();
-                            MessagingCenter.Send(errormessage, "LocationError");
-                        });
-                    }
-                }
-            }, token);
-        }
-    }
-}
