+++ Notify/Notify/Notify/ViewModels/WelcomePageViewModel.cs
@@ -16,6 +16,8 @@ namespace Notify.ViewModels
         #endregion
 
         #region Properties
+
+        private LocationService locationService;
         
         private string m_UserName;
         private string m_Password;
@@ -53,7 +55,7 @@ namespace Notify.ViewModels
             bool debugAutoLogin = true;
             IsBusy = true;
             
-            manageLocationTracking();
+            locationService.ManageLocationTracking();
 
             if (!debugAutoLogin)
             {
@@ -86,60 +88,9 @@ namespace Notify.ViewModels
 
         #region Private Functionality
         
-        private void current_PositionChanged(object sender, Plugin.Geolocator.Abstractions.PositionEventArgs e)
-        {
-            Console.WriteLine($"{e.Position.Latitude}, {e.Position.Longitude}, {e.Position.Timestamp.TimeOfDay}");
-        }
-        
-        private async Task manageLocationTracking()
-        {
-            var permission = await Permissions.RequestAsync<Permissions.LocationAlways>();
-
-            if (permission == PermissionStatus.Denied)
-            {
-                // TODO Let the user know they need to accept
-                return;
-            }
-
-            if (Device.RuntimePlatform == Device.iOS)
-            {
-                if (CrossGeolocator.Current.IsListening)
-                {
-                    await CrossGeolocator.Current.StopListeningAsync();
-                    CrossGeolocator.Current.PositionChanged -= current_PositionChanged;
-
-                    return;
-                }
-
-                await CrossGeolocator.Current.StartListeningAsync(TimeSpan.FromSeconds(1), 10, false,
-                    new Plugin.Geolocator.Abstractions.ListenerSettings
-                    {
-                        ActivityType = Plugin.Geolocator.Abstractions.ActivityType.AutomotiveNavigation,
-                        AllowBackgroundUpdates = true,
-                        DeferLocationUpdates = false,
-                        DeferralDistanceMeters = 10,
-                        DeferralTime = TimeSpan.FromSeconds(5),
-                        ListenForSignificantChanges = true,
-                        PauseLocationUpdatesAutomatically = true
-                    });
-
-                CrossGeolocator.Current.PositionChanged += current_PositionChanged;
-            }
-            else if (Device.RuntimePlatform == Device.Android)
-            {
-                if (Preferences.Get("LocationServiceRunning", false) == false)
-                {
-                    startService();
-                }
-                else
-                {
-                    stopService();
-                }
-            }
-        }
-
         private async Task initialize()
         {
+            locationService = new LocationService();
             VersionTracking.Track();
             if (VersionTracking.IsFirstLaunchEver)
             {
@@ -150,57 +101,9 @@ namespace Notify.ViewModels
                 await Shell.Current.GoToAsync("///welcome");
             }
             
-            subscribeToLocationMessaging();
+            locationService.SubscribeToLocationMessaging();
         }
-
-        private void subscribeToLocationMessaging()
-        {
-            if (Device.RuntimePlatform == Device.Android)
-            {
-                MessagingCenter.Subscribe<LocationMessage>(this, "Location",
-                    message =>
-                    {
-                        Device.BeginInvokeOnMainThread(() =>
-                        {
-                            Console.WriteLine(
-                                $"{message.Latitude}, {message.Longitude}, {DateTime.Now.ToLongTimeString()}");
-                        });
-                    });
-
-                MessagingCenter.Subscribe<StopServiceMessage>(this, "ServiceStopped",
-                    message =>
-                    {
-                        Device.BeginInvokeOnMainThread(() => { Console.WriteLine("Location Service has been stopped!"); });
-                    });
-
-                MessagingCenter.Subscribe<LocationErrorMessage>(this, "LocationError",
-                    message =>
-                    {
-                        Device.BeginInvokeOnMainThread(() => { Console.WriteLine("There was an error updating location!"); });
-                    });
-
-                if (Preferences.Get("LocationServiceRunning", false) == true)
-                {
-                    startService();
-                }
-            }
-        }
-
-        private void startService()
-        {
-            var startServiceMessage = new StartServiceMessage();
-            MessagingCenter.Send(startServiceMessage, "ServiceStarted");
-            Preferences.Set("LocationServiceRunning", true);
-            Console.WriteLine("Location Service has been started!");
-        }
-
-        private void stopService()
-        {
-            var stopServiceMessage = new StopServiceMessage();
-            MessagingCenter.Send(stopServiceMessage, "ServiceStopped");
-            Preferences.Set("LocationServiceRunning", false);
-        }
-
+        
         #endregion
     }
 }
